requirements.txt
invisible-watermark==0.2.0
numpy==1.24.0
scipy==1.14.1
pandas==1.5.3
opencv-python==4.5.5.62
Pillow==9.1.0
matplotlib==3.6.3
plotly==5.3.0
torch==2.5.1+cpu
torchvision==0.20.1
timm
open_clip_torch
gradio==2.2.15
fastapi==0.111.0
uvicorn==0.19.0
starlette==0.37.2
tqdm==4.64.0
imWatermark